requirements.txt
crewai[tools]
openai
chromadb
ragas
datasets
pandas
sqlalchemy
python-dotenv
tiktoken
fastapi
uvicorn
litellm